dashboard/requirements.txt
streamlit
google-cloud-bigquery[pandas]
pandas
